Add files via upload
This version of the proxyDrawsMSB function is similar to the other proxyDrawsMSB function, but takes as inputs the *summary statistics* for the non-selected cases (mean, variance) instead of the microdata. It *only* takes summary stats for non-selected; the other version would be used if you have microdata. I realized that I gave the functions both the same name; perhaps it would be worth changing one at some point (or combining into one function).